# apps/reference/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Iris reference application",
    "description": "",
    "version": "0.1.0"
  },
  "paths": {
    "/api/auth/callback": {
      "get": {
        "operationId": "callback",
        "responses": {
          "303": {
            "description": "Login complete"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "login",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginInfo"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "BrowserSession": []
          }
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "operationId": "logout",
        "responses": {
          "204": {
            "description": "This session logged out"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "BrowserSession": []
          }
        ]
      }
    },
    "/api/auth/session": {
      "get": {
        "operationId": "session_info",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionInfo"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorCode": {
        "type": "string",
        "enum": [
          "csrf",
          "login_failed",
          "unauthorized",
          "internal"
        ]
      },
      "LoginInfo": {
        "type": "object",
        "required": [
          "authorization_url"
        ],
        "properties": {
          "authorization_url": {
            "type": "string"
          }
        }
      },
      "Problem": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "SessionInfo": {
        "type": "object",
        "required": [
          "csrf_token"
        ],
        "properties": {
          "csrf_token": {
            "type": "string"
          },
          "user_id": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      }
    },
    "securitySchemes": {
      "BrowserSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-iris-session",
        "description": "Same-origin session and X-Iris-Csrf required. Explicit HTTP-loopback test mode uses iris-session-dev."
      }
    }
  }
}